portfolio ready

+++ frontend/src/screen/adminScreens/ClientsList.js
@@ -8,15 +8,12 @@ import { deleteUser, listUsers } from '../../actions/userActions';
 
 const ClientsList = ({history}) => {
 
-    const [open, setOpen] = useState(false);
   
-  
-
   const dispatch = useDispatch()
 
   const userList = useSelector((state) => state.userList)
   const {loading,error,users} = userList
-  console.log(users);
+//   console.log(users);
   
   const userDelete = useSelector((state) => state.userDelete)
   const { success:successDelete} = userDelete
@@ -33,16 +30,17 @@ const ClientsList = ({history}) => {
       dispatch(listUsers())
     }
     else{
-    history.push("/");
+    history.push("/client");
     }
   }, [dispatch,history,userInfo,successDelete])
 
-  const onOpenModal = () => setOpen(true);
-  const onCloseModal = () => setOpen(false);
- 
+
   const deleteHandler = (id) => {
-        dispatch(deleteUser(id))
-    }
+    if (window.confirm("Are you sure")) {
+        dispatch(deleteUser(id));
+      }
+    
+  };
 
     return (
         <div className="bg-primary-50 pb-24">
@@ -99,16 +97,20 @@ const ClientsList = ({history}) => {
                         <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                         <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z" />
                         </svg>
-                        </Link>
-
+                        </Link >
+                        {!client.isAdmin && 
+                        <button onClick={() => deleteHandler(client._id)}>
                         <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6 text-red-700" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                         <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
                         </svg>
+                        </button>
+                            }
+                        
                         </td>
                       
-                        
+                       
                     </tr>
-
+                        
                     ))}
                    
                 </tbody>
